Bugfixes - Removed __destruct from ControllerBase. - Model: added without functionality to prevent items from being output. - Model: changed order of invokeElement so \Closures have priority over relations. - Model: fixed issue in toArray causing null values to be excluded when using the filter param. - ModelCollection: simplified the toArray method. - TaglibRenderer trait: optimisations. - PhtmlNode: fixed deprication warning. - TaglibJs: changed tagTemplate to tagContainer. - WidgetTaglib: class is deprecated - use TaglibRenderer trait.
